requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
Pillow==10.1.0
playwright==1.40.0
PyJWT==2.8.0
gunicorn==21.2.0
Werkzeug==3.1.3